Created own state slice files for counter and authentication.
