+++ src/components/Auth.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import classes from './Auth.module.css'
 import { useDispatch } from 'react-redux'
-import { authActions } from '../store'
+import { authActions } from '../store/auth-slice'
 
 const Auth = () => {
   const dispatch = useDispatch()
