+++ src/components/Header.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import classes from './Header.module.css'
 import { useDispatch, useSelector } from 'react-redux'
-import { authActions } from '../store'
+import { authActions } from '../store/auth-slice'
 
 const Header = () => {
   const dispatch = useDispatch()
