+++ src/store/counter.js
@@ -1,0 +1,23 @@
+import { createSlice } from '@reduxjs/toolkit'
+const initialCounterState = { counter: 0, showCounter: true }
+const counterSlice = createSlice({
+  name: 'counter',
+  initialState: initialCounterState,
+  reducers: {
+    increment (state) {
+      state.counter++
+    },
+    increase (state, action) {
+      state.counter += action.payload
+    },
+    decrement (state) {
+      state.counter--
+    },
+    toggle (state) {
+      state.showCounter = !state.showCounter
+    }
+  }
+})
+
+export const counterActions = counterSlice.actions
+export const counterReducer = counterSlice.reducer
