+++ src/store/index.js
@@ -1,39 +1,9 @@
-import { configureStore, createSlice } from '@reduxjs/toolkit'
-
-const initialCounterState = { counter: 0, showCounter: true }
-const counterSlice = createSlice({
-  name: 'counter',
-  initialState: initialCounterState,
-  reducers: {
-    increment (state) {
-      state.counter++
-    },
-    increase (state, action) {
-      state.counter += action.payload
-    },
-    decrement (state) {
-      state.counter--
-    },
-    toggle (state) {
-      state.showCounter = !state.showCounter
-    }
-  }
-})
-
-const initialAuthState = { isAuthenticated: false }
-const authSlice = createSlice({
-  name: 'authentication',
-  initialState: initialAuthState,
-  reducers: {
-    logIn (state) { state.isAuthenticated = true },
-    logOut (state) { state.isAuthenticated = false }
-  }
-})
+import { configureStore } from '@reduxjs/toolkit'
+import { counterReducer } from './counter'
+import { authReducer } from './auth-slice'
 
 const store = configureStore({
-  reducer: { counter: counterSlice.reducer, auth: authSlice.reducer }
+  reducer: { counter: counterReducer, auth: authReducer }
 })
 
-export const counterActions = counterSlice.actions
-export const authActions = authSlice.actions
 export default store
